Move sample task to compiled AddTask
